requirements.txt
Flask
gunicorn
requests
mysql-connector-python
